Avoid using database/sql for ClickHouse NATIVE protocol
