+++ app/server/datasource/rdbms/clickhouse/connection_manager.go
@@ -2,84 +2,23 @@ package clickhouse
 
 import (
 	"context"
-	"crypto/tls"
-	"database/sql"
 	"fmt"
-	"time"
 
-	"github.com/ClickHouse/clickhouse-go/v2"
-	"github.com/ydb-platform/ydb-go-genproto/protos/Ydb"
 	"go.uber.org/zap"
 
 	api_common "github.com/ydb-platform/fq-connector-go/api/common"
-	"github.com/ydb-platform/fq-connector-go/app/server/conversion"
 	rdbms_utils "github.com/ydb-platform/fq-connector-go/app/server/datasource/rdbms/utils"
-	"github.com/ydb-platform/fq-connector-go/app/server/paging"
 	"github.com/ydb-platform/fq-connector-go/common"
 )
-
-var _ rdbms_utils.Connection = (*Connection)(nil)
-
-type Connection struct {
-	*sql.DB
-	logger common.QueryLogger
-}
-
-var _ rdbms_utils.Rows = (*rows)(nil)
-
-type rows struct {
-	*sql.Rows
-}
-
-func (r rows) MakeTransformer(ydbTypes []*Ydb.Type, cc conversion.Collection) (paging.RowTransformer[any], error) {
-	columns, err := r.ColumnTypes()
-	if err != nil {
-		return nil, fmt.Errorf("column types: %w", err)
-	}
-
-	typeNames := make([]string, 0, len(columns))
-	for _, column := range columns {
-		typeNames = append(typeNames, column.DatabaseTypeName())
-	}
-
-	transformer, err := transformerFromSQLTypes(typeNames, ydbTypes, cc)
-	if err != nil {
-		return nil, fmt.Errorf("transformer from sql types: %w", err)
-	}
-
-	return transformer, nil
-}
-
-func (c Connection) Query(ctx context.Context, query string, args ...any) (rdbms_utils.Rows, error) {
-	c.logger.Dump(query, args...)
-
-	out, err := c.DB.QueryContext(ctx, query, args...)
-	if err != nil {
-		return nil, fmt.Errorf("query context: %w", err)
-	}
-
-	if err := out.Err(); err != nil {
-		defer func() {
-			if closeErr := out.Close(); closeErr != nil {
-				c.logger.Error("close rows", zap.Error(closeErr))
-			}
-		}()
-
-		return nil, fmt.Errorf("rows err: %w", err)
-	}
-
-	return rows{Rows: out}, nil
-}
 
 var _ rdbms_utils.ConnectionManager = (*connectionManager)(nil)
 
 type connectionManager struct {
 	rdbms_utils.ConnectionManagerBase
-	// TODO: cache of connections, remove unused connections with TTL
 }
 
 func (c *connectionManager) Make(
-	_ context.Context,
+	ctx context.Context,
 	logger *zap.Logger,
 	dsi *api_common.TDataSourceInstance,
 ) (rdbms_utils.Connection, error) {
@@ -87,60 +26,15 @@ func (c *connectionManager) Make(
 		return nil, fmt.Errorf("currently only basic auth is supported")
 	}
 
-	var protocol clickhouse.Protocol
-
 	switch dsi.Protocol {
 	case api_common.EProtocol_NATIVE:
-		protocol = clickhouse.Native
+		return makeConnectionNative(ctx, logger, dsi, c.QueryLoggerFactory.Make(logger))
 	case api_common.EProtocol_HTTP:
-		protocol = clickhouse.HTTP
+		return makeConnectionHTTP(ctx, logger, dsi, c.QueryLoggerFactory.Make(logger))
 	default:
 		return nil, fmt.Errorf("can not run ClickHouse connection with protocol '%v'", dsi.Protocol)
 	}
 
-	opts := &clickhouse.Options{
-		Addr: []string{common.EndpointToString(dsi.GetEndpoint())},
-		Auth: clickhouse.Auth{
-			Database: dsi.Database,
-			Username: dsi.Credentials.GetBasic().Username,
-			Password: dsi.Credentials.GetBasic().Password,
-		},
-		// Set this field to true if you want to see ClickHouse driver's debug output
-		Debug: false,
-		Debugf: func(format string, v ...any) {
-			logger.Debug(format, zap.Any("args", v))
-		},
-		// TODO: make it configurable via Connector API
-		Compression: &clickhouse.Compression{
-			Method: clickhouse.CompressionLZ4,
-		},
-		Protocol: protocol,
-	}
-
-	if dsi.UseTls {
-		opts.TLS = &tls.Config{
-			InsecureSkipVerify: false,
-		}
-	}
-
-	conn := clickhouse.OpenDB(opts)
-	if err := conn.Ping(); err != nil {
-		return nil, fmt.Errorf("conn ping: %w", err)
-	}
-
-	const (
-		maxIdleConns    = 5
-		maxOpenConns    = 10
-		connMaxLifetime = time.Hour
-	)
-
-	conn.SetMaxIdleConns(maxIdleConns)
-	conn.SetMaxOpenConns(maxOpenConns)
-	conn.SetConnMaxLifetime(connMaxLifetime)
-
-	queryLogger := c.QueryLoggerFactory.Make(logger)
-
-	return &Connection{DB: conn, logger: queryLogger}, nil
 }
 
 func (*connectionManager) Release(logger *zap.Logger, conn rdbms_utils.Connection) {
