+++ scripts/debug/docker-compose.yaml
@@ -3,12 +3,12 @@ version: '3'
 services:
 
   clickhouse:
-    image: clickhouse/clickhouse-server
+    image: clickhouse/clickhouse-server:23.8
     container_name: yq-connector-clickhouse
-    ports:
-      - '8123:8123'
-      - '9000:9000'
-      - '9009:9009'
+      #ports:
+      #  - '8123:8123'
+      #  - '9000:9000'
+      #  - '9009:9009'
     environment:
       CLICKHOUSE_DB: dqrun
       CLICKHOUSE_USER: admin
@@ -18,7 +18,8 @@ services:
       - /tmp/yq-connector/clickhouse/data:/var/lib/clickhouse
       - /tmp/yq-connector/clickhouse/logs:/var/log/clickhouse-server
       - ./clickhouse/init:/docker-entrypoint-initdb.d
-      - ./clickhouse/config/z_log_disable.xml:/etc/clickhouse-server/config.d/z_log_disable.xml
+      # - ./clickhouse/config/z_log_disable.xml:/etc/clickhouse-server/config.d/z_log_disable.xml
+    network_mode: "host"
     ulimits:
       nproc: 65535
       nofile:
@@ -27,7 +28,7 @@ services:
 
   postgresql:
     image: postgres
-    container_name: yq-connector-postgresql
+    container_name: fq-connector-postgresql
     ports:
       - '5432:5432'
     environment:
@@ -41,7 +42,7 @@ services:
 
   minio:
     image: minio/minio:latest
-    container_name: yq-connector-minio
+    container_name: fq-connector-minio
     command: server --console-address ':9001' /data/
     ports:
       - '29000:9000'
